Fix several issues with ImageList.
One is bug 1933327

It seems to be timing issue. We assign the value of mAllImages in only
"onCreate" and will set it to null in "onActivityResult".
I modified the code so that it won't be null after "onCreate()".

The other issue is the crop image won't . The reason is we don't clean
up the cache when we reopen the cache. It won't work for ImageListUber
for the original fix since "buildImageListFromUri" won't return ImageList
of type ImageListUber.

The final one is the "getImaegForUri", last time, I made a change but
cause review fails in Camera. But, I need the check if the given uri is
actually the same as the image list not just check the "id".