Fixed the OOM in crop image.
The original code of "CropImage" won't wait the save action to be complete
before exiting.

We have to hide the ProgressDialog in onStop(), otherwise, if the job is done
in background(after onStop()), a "no window token exception" will be
thrown when dialog.dismiss() is called. Thus, a new MonitoredActivity is
introduced to help us handle this, hope it will be useful in other places.