Fix bug 1970161.
I think the reason is that we post a runner to do clean up when the
background job has been done.

However, the runner could be queued when the activity finished().
So, we first see a leak window message (because we havn't close the dialog yet)
And then see a "view not attached" exception (because we try to close
it after the activity being destroyed)

Fixed now.