Open Google Maps by default for "Show on map". DO NOT MERGE
This will solve an issue that browser will crash when open the http URL for map.
(bug 2039537). But we need to use these kind of intent for GMM to show a marker
on map. So, we change the code to open it with GMM first. If GMM is not installed,
we will use others who can handle the "geo:lat,lng" intent.