Fix bug 2078592.
This bug is that the touch event location should be traslated to be related
to the "owner view" of the zoom control. However, we didn't do that. So, there
is a shift of the location in the event object.

We tried to detect if user has touched on the zoom buttons and then keep those
on-screen controls (including zoom buttons) shown. Now, we fix this bug by
changing the logic to "if user has touched the screen while those on-screen
controls are visilbe" then we keep those controls on screen.

Change-Id: I3664c47d308a76afc7d8746da8bbbc1f3afbf7ac