Fix 2122382: We don't handle overlay destroy and re-create properly.
Because the overlay behavior changed, the overlay will be destroyed if we hide it.
So we need to stop preview at that time. To properly handle this situation we need
to add more code and it will be slower. For now we just don't hide the overlay
because the image on top of it will obscure it anyway.

Change-Id: I5c479381a578448a712dedc1bb97ba6f36db468f