Temporally solve the overlay creation problem.
Fix bug:  http://b/issue?id=2153980

Because of bugs of overlay and layout, we sometimes will try to layout the
viewfinder in the portrait orientation and thus get the wrong size of
mSurfaceView. When we cange the size of preview size the new overlay will be
created before the old one closed, and cause an exception. For now, just get
the screen size

Change-Id: I3daf4bce801507aaaca1f030f09d6e18cb23f0f2