Remove use of Settings.Secure.USE_LOCATION_FOR_SERVICES
This Google-specific setting is going away because of unbundling.

This change will cause geotagging to be off by default, instead of
being set by SetupWizard by the "Use location for Google services"
prompt. If this is not desirable, another prompt could possibly be
added to SetupWizard, or a new framework geotagging setting
could be added and SetupWizard be modified to set that as part
of the Google location opt-in.

Part of: http://b/issue?id=2383870

Change-Id: Ic160c2260dd597a35d79fec93250a816fbee8b81