Camera: Fix Camera Crash issue in Video-Recording seen when sdcard is mounted after the camcorder is initialized.
When sdcard is inserted after the camcorder is started, the videofilename is not intialized as the sdcard path is not accessible.
Media-Recorder intialization is called for after the sdcard is inserted without releasing it.
As a result the initializerecorder returns as it finds the mediarecorder not null.
Re-initializing the mediarecorder from scratch, when the sdcard is mounted fixes the issue.

Change-Id: Ie9790ba74eb38f2b1e5b0d0747c8db73faae57b9