Fix the ANR after switching between Camera and Camcorder.
The reason of this bug is we try to queue an event in the GLThread, but the thread
has already dead when the GLRootView detacched from the window. So, we are
waiting a task that will never be executed.

I have seen a similar ANR in GLRootView.onTouch(), so I also fix the that issue
in this change.

Bug: 2559472

Change-Id: I49efd9ca01f1f6cce73320c31448ebaa7687469f