Nudge the camera preview 2dip from the screen edge.
On Sapphire, the preview background was right up against the
edge of the screen; since that portion of
border_view_finder.9.png is very dark, it looked as if there
was no border at all. Now there's a guaranteed 2dip of
bg_camera_pattern.png texture visible to the left of the
viewscreen.

Bug: 2655309
Change-Id: I4d4f45542700b8b0fec564b72895e42d9dd29da3