Fix wrong orientation after suspend and resume.
Camera application calls setDisplayOrientation according to
Display.getRotation and sensor orientation. After power key is
pressed, camera app resumes behind the lock screen and preview
is started. Rotation is 90 at the time. After the screen is
unlocked, surfaceChanged is called because the orientation is
changed from portrait to landscape. Preview is stopped and
restarted. setDisplayOrientation should be called again because
the rotation has changed to 0.

bug:2870671
Change-Id: I1ca5106087c168d88cc8eb4f2a9f192d7fa75a8a