Set audio input source type to CAMCORDER in camcorder app.
Up to Froyo release, the media framework (pv opencore in particular)
was forcing the audio input source type to CAMCORDER whenever the
capture was for video and audio. This is not true anymore on
gingerbread with stagefright. The proposed solution is to specify
the audio input source as CAMCORDER in the camera app, and not
enforce it in the framework. Enforcing it in the framework removes
flexibility like recording from the front camera with a mic different
from the back mic on platforms with 2 mics and 2 cams.

bug:3171482
Change-Id: I5024ebb453054bfc0fcafedcdc78d98b0393bf81