Camera: Make sure hardware is unlocked before starting encoder
There's a race condition between the surface management and the
encoder. With the corrent code, if the preview buffers aren't
ready before the encoder is started, the app force-closes when
the resolution is changed. On slower devices (or simply slower
camera userspace), this is 100% of the time.

This makes sure the hardware isn't locked by the camcorder before
starting the encoder. It isn't the ideal solution, but it appears
to work.

Seanzer: I've removed some redundant unlock code that was causing the Camera.apk to fc when the screen would go off and on.

Change-Id: Icab2e651ad58a317b82c09a9b8f64a54a54705d5