Hide the next/previous button if the setting is the last or the first.
Other settings use next and previous button to control the settings.
If the setting is the first one, do not show the previous button.
If the setting is the last one, do not show the next button.

Also fix a bug that sometimes the state of long press is wrong.

bug:3273488
Change-Id: I45d6a5035c0175c8f1808c46078b0ff2291209cc