camera: Allow zoom and parameter changes during video recording.
Previously, it was not possible in the Android framework to interact
with the camera after MediaRecorder.setCamera() was called. With the
recent changes by Qualcomm, we can now adjust all parameters while
recording is in progress.

Zoom functionality has also been added to the camcorder, however it is
disabled in 720p mode. This appears to be a driver limitation.

Change-Id: I5f001518f9840a82e06919957bee1547501d3bbc