Fix wrong preview orientation when device is held upside down.
Suppose a user opens camera app and press home to exit.
The user holds the device upside down and opens camera again.
The animation takes some time so camera app will get the old
rotation value in onResume and surfaceChanged. Besides, framework
does not have a callback for it. The solution is to poll the
rotation value for 5 seconds. If it changes, restart the preview.

bug:3362860

Change-Id: I065e8601e321361d8db405f9c587ec2a5bce050c