Build-time option to support cameras with counterclockwise rotation
There's at least one camera (the P990's) that needs this: the rotation
direction is the opposite of what it should be, so the "upright phone"
portrait orientation is 270 degrees, and upside-down is 90, which
results in upside-down portrait snapshots.
The landscape orientations are fine, since the camera "natural position"
is still the standard landscape on the left side.

Change-Id: I2ed64ddda95bcaac93a65d812cde7eda2776978d