Merge the indicator bar implementations.
bug:5039235

Originally we have the GLView-implemented indicators for Phone-based UI.
Based on the current UI design, we would like to merge the look on all
devices for the dialog of the camera setting. Therefore, a refactor of
how to handle the popup window is required here and this will simplify our
UI work for handling different layout/resolution.

Change-Id: I6f9e6ebed7600be460ddd906505cdc44f81098ad