Camera: Make sure flip is reset on nvidia cameras
Flip can be used to correct the output (dataCallback) frames used
by other apps, but in here it breaks things due to the inability
to flip the overlay.

Apps like CameraFun and GoogleTalk need that reversal, and compensate
for it in different ways (talk by resetting and reorienting the overlay,
CameraFun I have no idea) and appear to be working fine.

Change-Id: Ib20d2a9f06dd9943dcc5ce2a8ccb26159a4686e7