Support camcorder mode for galaxys2.
A new resource is added to the Camera app to enable the proper method
for starting video recorder mode on galaxys2. This phone must set
camera parameter cam_mode=1 to enable the 1280x720 and 1920x1080
preview resolutions, but can not be detected from parameters as this
parameter is unset by default. If other devices need to use this
method, they can set needsSamsungCamMode to true in their overlay.

Change-Id: I629ed74025aaf1f4055a4e910101ed252143d021