Added matrix metering support
While Samsung is the only brand of phone I know with support for
matrix metering, these values are very generic and I believe they
can be added to the Camera app. If they are to be added to overlays,
they will need to be duplicated in at least two overlays and
localization will be very hard.

Change-Id: I70f755a1bceaff868c7e4dde1f716fa3866c8c36