Fix an issue with hitting frame number limit.
In the native side, the library always returns the transformation of the 100th
frame after hitting the frame number limit - 100. The application codes
need to avoid adding more frames once the limit reaches. Otherwise, we
see the UI repeating the 100th on the panorama preview. Ultimately, we
want this termination condition to be solely determined in the native
side.

Also updates a comment on MosaicRenderer.

Change-Id: I6883b25f14103527de8454a2568f4768669e0aff