Fix AF failure when shutter and tap focus are triggered at the same time.
When the shutter button and the viewfinder are tapped simultaneously.
(1) onTouch ACTION_DOWN. Tap area is set and AF is not started yet.
(2) Shutter doFocus. AF starts.
(3) onTouch ACTION_UP. AF is called again and an exception is thrown.

Now onTouch always cancelAutoFocus if AF is in progress and the state
is not STATE_FOCUSING_SNAP_ON_FINISH.

bug:5265128
Change-Id: Ib52dcc5169132dde2b688062d812d59e0d7c2d7c