Camcorder profile: cleanup and prepare custom media profiles
This commit was a part of the WIDE camcorder media profile add-on
Devices are very different, and the WIDE profile is not enough generic.

Actually, it is really hard to add a new media profile in frameworks/base,
the change will allow, later, to use a new property to set number of profiles.

Change-Id: I91d12984d1480da6744f4cf9862a6a08b588ef72