Camcorder: Fall back to lowest quality if media profile is wrong
This commit is a safe part of the WIDE camcorder media profile add-on
Devices are very different, and the WIDE profile is not enough generic.

This change prevent also fc when one of the media profiles is wrong,
falling back to the "low" quality mode with a detailed error in logcat.

Actually, it is really hard to add a new media profile in frameworks/base,
the change will allow, later, to use a new property to set number of profiles.