Fix display of preview thumbnail after recording video.
The first was that the code was trying to rotate the video thumbnail
retrieved from MetadataRetriever back by just reversing whatever
rotation we set as a playback hint when we started recording. But
this isn't actually correct -- the bitmap we get back from MetadataRetriever
is going to be oriented correctly for display if the device is oriented
in the locked UI orientation (landscape), and we need to rotate it
to compensate for what the actual device orientation is relative to
that.

Phrased another way, we were taking into account the camera and
device orientation, but not the UI orientation. This just happened
to work on some devices because of their camera orientations, but not all.

bug:5360349

Change-Id: I8b481907c211328726ecd91fa054b9e9a4798601