Reduce the frame memory we allocate.
We increment the LR frame counter and the HR frame counter at the same
time and check the frame limit condition based on which counter hits the
limit first. We can just combine the MAX_FRAMES_HR and MAX_FRAMES_LR
to one. This reduces the memory we allocate to LR frames by 8~20 MBytes
depending on the size of an input frame.

Change-Id: I1b76d94db09985df21c946830295b94f39e99e59