bugfix 5422126 Panorama needs to load the thumbnail from the file
This change makes Panorama closer to Camera in the way the Thumbnail is
loaded.  But the Camera has a pretty different startup sequence so you
should review this to make sure it's correct.

Note that doing any IO on the UI thread can lead to ANR so we should
consider making the thumbnail load from an async task instead.

NOTES:

In camera the sequence is like this:

Thumbnail.loadFrom
initThumbnailButton
initializeFirstTime
FIRST_TIME_INIT message
onResume, surfaceChanged,

So we initiate the call to Thumbnail.loadFrom via onResume or surfaceChanged.

I removed the call to "updateThumbnailButton" from onStart to make this closer
to what we do in Camera.

Change-Id: I86283acbc9355f077d2966a018a3b972efe1c08c