Handle MediaRecorder stop exception for effects recording
When recording videos with the effects ON, if user starts and stops the
video in rapid succession, MediaRecorder stop can throw an exception,
which we want to be treated separately than other runtime exceptions
thrown by effectsrecorder.
In this particular case, the idea is to keep the app running without crashing.

Currently, all "Stop" exceptions will lead to the same behavior.
In future, we might need to work out the details more to find out
what caused the stop to crash and pass that up from the native layer
and react according to the cause of the exception.

Change-Id: I35a961d6cf4fc3cc3ca66870ea80ee16316f0728