More bg training polish.
Make training modal by placing a semi-transparent view behind the
message panel to prevent interaction with the rest of the camera UI.

Use the framework dialog asset as the background asset for the message
panel. Since this asset is not public I had to copy it into the Camera
package.

bug:5225962
Change-Id: Ief97f545c6c0bb3695a056481d5e6d56d44edf25