bugfix 5479807 Video preview displaced to right when orientation changed
Three changes:

(1) center the play button over the thumbnail
(2) leverage the existing mechanism for scaling the image for video
thumbnails which want to be 4/3
(3) use center_crop for video previews so that they properly align
in the ui.  Not doing so means we were getting black bars at the top
and bottom which was wrong.

Change-Id: Ifc416455d03136a393239487db07493dc27f151e