Effects: shut down camera preview when unexpected graph closures occur.
If the graph runner needs to shut down a graph due to an error, previously
the surface texture connecting the camera to the processing graph would get
released before camera was told to shut down preview. This led to various
abandoned surfacetexture errors in the camera HAL, sometimes including
native crashes.

Now properly uses the SurfaceTextureSource callback to detect unexpected
graph shutdown to clean up the camera before the SurfaceTexture is released.

Bug: 5451833
Change-Id: Id130c1e5deb27cffab335affaa1ed55f04c260a5