Add missing state check in EffectsRecorder startup path.
When starting up, the EffectsRecorder gets a callback when the effect
filter graph provides an input SurfaceTexture. This callback was not
checking the EffectsRecorder state correctly, and if stopPreview was
called on the EffectsRecorder between startup and before the callback
fires, the callback would improperly reset the recorder state,
resulting in NPEs down the line.

Bug: 5493458
Change-Id: If552b7c0ea82e1506f0b22eb5fa0447c1ca4adf8