Reset mCurrentVideoUri to null before starting recording.
While running monkey, we encountered a usecase where Video Recording is inProgress and an
intent to launch Panorama Activity came in. While switching StopVideoRecording failed and
resulted in Camera Exception (observed 3 times in 2 weeks of testing). mCurrentVideoUri
is not cleared, so the thumbnail of the last video is re-generated, which can take 3
seconds.

bug:5596264

Change-Id: Ie14c426c81c46f2d9a6741b6213cf66b1dbb5612
Signed-off-by: Braganza Sam Roberts<sam.braganza@samsung.com>
Signed-off-by: Heechan Park <heech.park@samsung.com>