Fix crash after restoring defaults.
After restoring defaults in camcorder mode, the indicator
control calls onSharedPreferenceChanged if record location
is on. onSharedPreferenceChanged is also called in the end
of restorePreferences.

With two onSharedPreferenceChanged calls in a row, the first
call notices effects have been disabled, and begins the
effect pipeline teardown in the background thread. Normally,
this would conclude in a few hundred milliseconds, and would
then restart regular preview.

However, with the second call to onSharedPreferenceChanged
coming in right after the first, the method no longer sees
an effect transition in progress (since current and new
effect are equal to none now), and thus just ends up
restarting preview. That attempt at restart runs into the
crash because the main SurfaceView has not yet been released
by the background effects thread.

To fix this, InLineSettingSwitch should not invoke the
listener when updating UI.

bug:5679133

Change-Id: I06aca382105455585b633dc63f014a8f71ccadbe