Added option to use either removable SD card or internal memory
On the "Advanced options" screen, added the option "Use external SD card",
with description "Store photos and videos on external (removable) SD card"
is enabled if a phone has both internal memory and removable SD card, and
the external card is mounted and writable. Otherwise this option is
disabled.  Takes into account the Cyanogen "persist.sys.vold.switchexternal"
property to detect which of /mnt/sdcard and /mnt/emmc is the true removable
SD card and makes the Camera save photos and videos to /DCIM directory there.

The default is as per prior Cyanogen setting, i.e.  use the same storage
as the main programs and data storage, but now can be easily switched if
the user runs out of memory while recording photos and videos, or can keep
all programs and data on internal phone storage, while swapping SD cards
as needed for recording more photos and videos.

Change-Id: Ibef2304640dfd7c060927d4268a10af1c388500a