Resolving Backdropper heterogeneous background
On quick start-stop with Background replacement, there was an issue.
The runner would error out and the graph would restart. The preview
loaded faster than the background learner could finish learning,
and hence the user could start recording the video before the background
video loaded. The end result would be a video with some
frames that do not have the intended background.

Bug: 5623585
Change-Id: Idd24c6df7360267366bbd33b9d24b9b4439e55fb