Cut down unnecessary getParameters() calls.
Move mParameters upwards to ActivityBase.

Everytime when setParameters() is invoked, we use mParameters as
argument, thus mParameters is virtually the updated cache of camera
parameters, so we don't have to get that from camera everytime when
we need it.

There are cases that setParameters() will result in camera parameters
different from the argument passed in. For those cases, we call
getParameters() immediately after setParameters() to get the updated
value for mParameters. This logic remains unchanged in this CL.

bug:5954716
bug:5954391
Change-Id: I4681aed5ebdbf9ee5df1ed3e7aab92912302ee3b