Camera: Use popup settings instead of knobs
Control knobs are hard to use. It's hard to get to a specific
setting and you don't know the list of setting without going
through it.

Pop-up settings allow you to see every setting with easy overview
and selection. Additionally, it allows more text to be displayed,
fixing the display problem of storage setting.

Change-Id: I22e9a5b9f49741c734e7403e0f091175ff94be0b