Fix camera control disappear when restarted.
This changelist fixes NPE in Camera.onConfigurationChanged()and
VideoCamera.onConfigurationChanged() by removing mCameraDevice
operation from setDisplayOrientation() and initializeZoom(). But the
camera control disappears when camera app is restarted. It's because
ActivityBase.onFullScreenChanged(boolean) is invoked three times with
the parameter value in this order true, false, true. The second
invocation leaves a pending action withEndAction(mHideCameraAppView),
which is executed in the third invocation of onFullScreenChanged(),
causing the camera control being hidden. The fix is to add
withEndAction(null).

Another bug is filed for that onFullScreenChanged() being invoked
three times, which causes the camera control flashes.

Bug:6438113
Change-Id: I36305eabe5d7092d681550a1d5772ec790eb1cbb