Fix exception on interrupted activity w/ effects
Fix b/5873421.
With effects on, there was an exception if the videocamera activity was
interrupted. This was due to the fact that the
video filedescriptor was being closed in onpause() whereas the
effectsrecorder needed that till later. So moved that closing the file
descriptor to the callback received from effectsrecorder. This also
needs the closeCamera() to be split into two parts: closing the camera
itself and closing the effects. Closing the effects is shifted
to the callback, too. Else, the effects are stopped and released in
the closeCamera() call, thus leaving no way to not close the file
descriptor in onPause(). Additionally, a method disconnectCamera is
added so that the effectsrecorder can handle a released camera when
shutting down.
Also, added a method to let the filter graph know that the underlying
surfacetexture will be disconnected in onPause().

Change-Id: Ic6aada9f7e816b320bf292aafd453efe6dc71073