Fix 5914: Reset Camera to Defaults Crashes Camera App
Reseting the camera settings to default values results in a gets -1 index when try to parse current default
value ("Off") to a index value, causing a ArrayIndexOfBounds. "Off" string and his value "1" are
not sinchronized.

This changes adds a default value of 1 for the setting, instead of use the string value.
Also prevents a call of reloadPreference method of InLineSettingRestore items, because
this kind of setting doesn't have a real setting preference (causing a NPE)
Remove an unnecessary code for set default index value to 4, that causes that when the Camera is initialized
the saved value of the preference is changed with the default value. Now the preference has a
default value if no user setting was set.

Change-Id: If5d48654a02991df8d788fb1d03a073e709df921