Do not use API in level 11.
Util:
Copy the implementation of AlertDialog.Builder.setIconAttribute()
in the framework.

EffectSettingPopup:
Effect recording is only used in honeycomb and above, so we can simply
add TargetApi for method reloadPreference().

FaceView:
It is only used when face detection is supported, so we can simply add
TargetApi for the class.

InLineSettingXXX:
Move the implementation of onPopulateAccessibilityEvent() into
dispatchPopulateAccessibilityEvent().

Bug: 6927726
Change-Id: I06205914f246b31e37a0c5c549c4cae8d5b11303