Add readme for camera app make
Developers trying to build Camera app will now be able to see that the Camera and Gallery2 packages have been combined into a single app.

While this seems trivial to anyone actually working with the Camera sources, it can save some time to developers who aren't familiar with this change.

Change-Id: I812ecb03be10418f0108166890429949c8f142d4