Finish the activity when screen turns off in secure mode.
When camera is started from the secure lock screen and the
screen turns off, finish the activity. Otherwise the camera
will be still on top of the lock screen when the screen turns
on. This also makes sure pictures taken in the last use will
not be shown when the camera is started from the lock screen
again.

bug:7053266
Change-Id: Id3df2a7787e43d993f55ed70374d6722e4bfd040