Revert "Camera: Add pinch to zoom"
If well, pinch to zoom is really awesome and that stuff, camera already has zoom control on the right side and it's better to have AOSP implementation slide to gallery on the entire screen that just on the buttons. This may also break swipe to gallery on LDPI devices (and even MDPI ?), in wich case buttons area is really small.

This reverts commit ef0f9d8f0c3929665925a1ce19ec692a72fa7d99