Reduce latency when camera is started from secure lock screen.
When camera is started from secure lock screen for the first
time after screen on, the activity gets
onCreate->onResume->onPause->onResume. To reduce the latency,
keep the camera for a short time so it does not need to be
opened again.

The activity is started while the lock screen is shown, so it
must be started and paused because it is not visible. Then the
lock screen is removed and now it is visible so can be resumed
again. This is, as defined, how things are supposed to happen.

bug:7110256

Change-Id: I6a36aeffce32c5e17d267b6fe591fe8c65668b9c