Making back button exit camera
Bug: 7208910

The bug is that whenever createCameraScreenNail was called, a new activity state
was added to the activity stack. But the activity states never got removed from
the stack until onDestroy is called (i.e. the end of cameraActivity). When back
button is pressed, the top state in the stack got removed and since there
were more than one activity states in the stack, the new top state was resumed,
which is why back button didn't make camera exit.

Instead of always adding a new activity state to the stack, this CL only adds
the state when the stack is empty. Otherwise, it switches the top state to the new state.
It requires changing switchState function in StateManager.java from being only
accessible within com.android.gallery3d.app to public.

This is part 2 of 2 in fixing the bug.

Change-Id: I1809e0250c9560235d5d4933c2878491f643b357