Zero Shutter Lag Mode on Front Camera
Add a config to determine if the Front Camera requires Zero Shutter Lag Mode
Few Front Camera Sensors (Like the one on I9300) doesn't support ZSL Mode and they would require the preview to be stopped and started manually during each Snapshot. However for ZSL Mode, the preview is never stopped which works fine for the Back Camera but for the Front Camera, it causes the preview to freeze after taking the first Snap. Hence, the new Config 'enableZSLFrontCam' will determine if ZSL is required on the front camera and if not, the preview is restarted. By default the setting is turned off since most of the Front camera sensors are usually inferior and don't support such advanced features

Change-Id: Id143325be1b41d61fb1d4a2b4e0297d3d8356c47