WIP: Searching a fix for back key on camera
When you press back key on camera app (Galaxy S2), it flashes like a photo
was taken, I don't have the time right now to debug where's the root issue of
this, but atleast we know is in this part of the code.
For now this fixes it but crashes camera when you open it for second time.

Change-Id: I3509a8800dc40f411ed2589e958ba77d3e251f50