Fix "Too fast" warning is not displayed in panorama capture.
If the frame rate is too low and the camera motion is fast,
the frame to frame motion estimation may fail -- leading to
incorrect translation estimates. Originally "Too fast" is
showed when degrees per second exceeds the threshold. But
different devices have different frame rate. Also, if
degrees per second is low but there are not enough frames,
the stitched panorama is still blurry. Now "Too fast" is
showed when degrees per frame is too high. This is more
accurate.

bug:7165018

Change-Id: I082e9f612becb9aaf5f453694914fb24b7e6ffa9