Fix camera screen nail for swiping during capture
Bug: 7341774

Flash animation is triggered by shutter click. But in the case
where swipe happens immediately after shutter click, when draw()
is called to draw the animation, the camera is already swiped away,
which causes camera screen nail not being drawn correctly.

This CL checks for whether camera is still full screen before the
animation is drawn, and skips the animation for non-full screen
cases.

Change-Id: I98456042db4080404fd233aed66e0f5dbf8ce960