Ensure the whole camera controls are displayed
The measuredSize which defines the width of the camera controls
layout is calculated as follows:
measuredSize = (int) (longSideSize - shortSideSize / 3.0 * 4.0);

On TouchPad the following values are being used
measuredSize = (1024 - 720 / 3.0 * 4.0) = 1024 - 960 = 64
so the calculated size is too small and the camera controls are cut.

The minWidth in res/layout-sw600dp/camera_control.xml is defined to
260 and minHeight in res/layout/camera_control.xml to 76.
Make sure the measuredSize is not smaller than the minWidth/minHeight
from the layout.

Change-Id: If512295ef584dcf7b69b4df5e56744a106286038