Adding videoSnapshotInNotPausedMode option to resources values.
Those legacy devices that does NOT support taking snapshot images while
video is recording, should set this to False in overlay.
Newer devices that can take snapshots in the video recording,
should leave this untouched (true).
The part where single tap is handled
will use this flag and choose to check the mPaused variable
accordingly to this setting.

This could resolve the issue arising in the following
patch: http://review.cyanogenmod.org/#/c/29039/2
which affects some of the devices that doesnt support this
feature correctly.

Change-Id: If9cc013796d28959256da0e3a22ac1c7427df044