Add a GPS indicator to the camera
This will add a GPS indicator to the preview of the photo camera. It is placed as an overlay at the upper top corner of the photo camera preview.
It shows one of three icons depending on the GPS state: acquiring fix, GPS fix acquired and GPS disallowed in system settings.

The display of the icon is optional, it can be deactivated by a new setting "Show GPS icon" in the camera preferences.
The new setting will be set to off and disabled (shown grayed) when the "store location" setting is set to off.

The icon will only be shown in the preview of the photo module, not in the video or panorama module, and not in the gallery view.

Patch Set 1: use cm_strings.xml and cm_arrays.xml instead of strings.xml and arrays.xml

Drawables by blunden.

TODO: translations for the new settings item.

Change-Id: I0686ef4ee1fe9b34ac9beaaa2bd43555234f4f4d